requirements.txt
autopep8
coverage
flake8
pylint

django==3.0.3
djangorestframework==3.11.0
